исправления по замечаниям

+++ core/main-service/src/main/java/ru/practicum/ewm/main/controller/comment/params/FindAllCommentsParams.java
@@ -23,9 +23,9 @@ public class FindAllCommentsParams {
 
     LocalDateTime rangeEnd;
 
-    Long from;
+    int from;
 
-    Long size;
+    int size;
 
     String sort;
 }
